refactor(security): colocate baseUrl allowlist in base.ts
Move validation into base.ts and allow only production, nonprod-dev, and nonprod-qa API hosts.
